+++ constants/ethermint_extensions.go
@@ -2,6 +2,5 @@ package constants
 
 const (
 	EthermintExtensionOptionsEthereumTx  = "/ethermint.evm.v1.ExtensionOptionsEthereumTx"
-	EthermintExtensionOptionsWeb3Tx      = "/ethermint.types.v1.ExtensionOptionsWeb3Tx"
 	EthermintExtensionOptionDynamicFeeTx = "/ethermint.types.v1.ExtensionOptionDynamicFeeTx"
 )