+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 - (deps) [#148](https://github.com/EscanBE/evermint/pull/148) Bumps Cosmos-SDK v0.50, CometBFT v0.38, IBC v8
 - (feemarket) [#155](https://github.com/EscanBE/evermint/pull/155) Deprecate usage of `x/feemarket` `NoBaseFee`
+- (eip712) [#158](https://github.com/EscanBE/evermint/pull/158) Cleanup EIP-712, remove `ExtensionOptionsWeb3Tx` and the custom Cosmos AnteHandler
 
 # Cosmos-SDK v0.47